+++ docs/core/extensions/snippets/localization/example/MessageService.resx
@@ -117,9 +117,6 @@
   <resheader name="writer">
     <value>System.Resources.ResXResourceWriter, System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
   </resheader>
-  <data name="DinnerPriceFormat" xml:space="preserve">
-    <value>On {0:D} my dinner cost {1:C}.</value>
-  </data>
   <data name="GreetingMessage" xml:space="preserve">
     <value>Hi friends, the ".NET" developer community is excited to see you here!</value>
   </data>
