+++ docs/core/extensions/snippets/localization/example/ParameterizedMessageService.resx
@@ -117,7 +117,7 @@
   <resheader name="writer">
     <value>System.Resources.ResXResourceWriter, System.Windows.Forms, Version=4.0.0.0, Culture=neutral, PublicKeyToken=b77a5c561934e089</value>
   </resheader>
-  <data name="GreetingMessage" xml:space="preserve">
-    <value>Zdravo prijatelji, ".NET" developer zajednica je uzbuđena što vas vidi ovde!</value>
+  <data name="DinnerPriceFormat" xml:space="preserve">
+    <value>On {0:D} my dinner cost {1:C}.</value>
   </data>
 </root>
