+++ docs/core/extensions/snippets/localization/example/example.csproj
@@ -16,6 +16,9 @@
     <EmbeddedResource Update="MessageService.resx">
       <Generator></Generator>
     </EmbeddedResource>
+    <EmbeddedResource Update="ParameterizedMessageService.resx">
+      <Generator></Generator>
+    </EmbeddedResource>
   </ItemGroup>
 
 </Project>
